Fixed the database structure with Enrique's suggestions
